connection: add cut-off value to quota accounting
Instead of accounting all early messages to the first user exceeding the
quota, we now have a cut-off value. All messages below that value are
never accounted. This means, we no longer incorrectly account messages to
the wrong user. Instead, we allow up to MAX_MSGS_PER_USER of un-accounted
messages before we start accounting. This means, at most twice the amount
of messages can be sent from a single user to another.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>