queue: pin users while holding their messages
Imagine user A sends a quota-accounted message to B and exits before the
message is retrieved. If user C connects to the bus, it might get the same
user->idr value as A (as it is free now). If C talks to B, it will re-use
the quota-counter of user A, even though it's unrelated.

Fix this by pinning domain_user objects in the message queue. As long as a
message is accounted for in a user-quota, we need to keep the domain_user
object alive.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>