handle/connection: cache privilege levels
This introduces two cached privileges:
  handle->privileged
  conn->privileged

Both values are a boolean and describe whether a handle or connection is
privileged. Unlike our recent privilege checks, these values are cached
during open() time and then remembered.

For connections, conn->privileged is true if either the process had
CAP_IPC_OWNER in the init-user-namespace or has the same fsuser as the bus
owner.

For control-handles, handle->privileged is true only if the process had
CAP_IPC_OWNER in the init-user-namespace.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>