domain, endpoint: drop cdevs (ABI break)
Now that we have everything in place for kdbusfs, we can start
dropping support for cdevs. At the same time, let's drop the domain
stacking entirely, and the code that creates a root domain at module
load.

This changes the userspace perspective of kdbus. However, for code
that does _not_ deal with domain creation, all that is needed to do
is change the base path of the kdbus tree.

Signed-off-by: Daniel Mack <daniel@zonque.org>