node: make activated nodes own a reference themselves
This changes the node implementation to take a ref-count whenever a node
is activated, and drop it on deactivation.

This allows us to create nodes, link them and then activate them. If we
then drop our caller ref, we automatically cause the node to be dropped
when its parent is deactivated.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>