test, test-policy-ns: when receiving from multiple threads check if queue is full
When receiving from multiple threads in kdbus_recv_in_threads(), the
threads share the same connection queue, but due to scheduling decisions
that queue can be filled, this makes the thread sending packets to fail
with -ENOBUFS or -EXFULL which will cause further receiving threads to
fail with -ETIMEDOUT when they try to catchup or when they are scheduled,
at that time the sender has already failed.

So improve this by checking if we failed due to scheduling decisions and
packets have been lost, return success, this is not our fault.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>