util: fix hash-function
We must not use full_name_hash()! This hash function accesses strings "a
word at a time" (on most systems). This is fast and nice for qstr-namei
lookups, but dangerous if we don't guarantee proper string-end alignment.
In our case, strings might not be cleared or even allocated after the last
byte, so we might get incorrect values or even fault.

Fix this by using the full hash-function, which is slower, but correct.
This is also what kernfs and friends are doing.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>