match: rename the match db argument from db to mdb to not confuse lockdep
Currently the match and policy functions use the 'db' as an argument to
refer to their match db or policy db, and unfortunatly both functions use
the same variable name 'entries_rwlock' to refer to their read-write
lock too. I find that this may confuse lockdep output.

So change the 'db' into 'mdb' in the match code and the 'entries_rwlock'
into 'mdb_rwlock'. We can also change the 'db' argument name in policy
if we want.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>