node: fix missed entries on iterate() with hash-clashes
Imagine we have two entries, A and B, with the same hash and dir_iterate()
has 'A' set as next entry (ie, set as f->private_data). Now 'A' is
deactivated and dropped from the rb-tree before we enter dir_iterate().

Once we enter dir_iterate(), we try to acquire f->private_data, which is
set to 'A'. This will fail, as it is already deactivated. Therefore, we
call into kdbus_node_next_child() with prev=='A'. This will look for the
next node with a _higher_ hash than 'A'.

This means, we will never return 'B', as it has the same hash as 'A'. If
we changed kdbus_node_next_child() to look for an entry with an higher
_or_ _equal_ hash to 'prev', than we might incorrectly return an entry
previous to 'A' multiple times.

The correct fix here is to use the name of 'prev' as additional hint:
Instead of looking for an entry with an hash greater than 'prev', we look
for an entry with a name greater than 'prev'. As two entries can never
have the same name, this is guaranteed to return the next entry in a
reliable manner.

Fix node_find_closest_unlocked() to take a name as optional argument to
get more precise results. We still need to support the case where we only
get the hash value (for llseek). In those cases, we always return the
lowest entry with the same hash.
Note that llseek() on directories does not neccessarily return unique
offsets for all entries, so it's totally safe to always seek to the first
matching entry.
llseek() on directories is usually only used for FRONT / TAIL seeks,
anyway. Those work reliably for us, so we're safe.

Note that we're even better than kernfs here! So our behavior should be
fine.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>