node: fix deactivation race
If two threads deactivate a node in parallel (like closing a bus and
endpoint at the same time) before the node was ever activated, we must
make sure not to drop any activation-reference more than once.

That is, we currently increase the ref-count by 1 on activation and drop
it on deactivation. If a node never got activated, we shouldn't try to
drop this reference on deactivation, but deactivation should still work.

Make sure to only drop this reference if the node as active and we're the
one task who actually added BIAS to mark in deactive.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>