node: acquire parent during kdbus_node_link()
We must make sure the parent is active when adding new links. Usually, we
require the callers to guarantee that, but it is easier if we verify this
ourself.

In case the parent is not active and we add the node, no-one guarantees
that the child-node is ever deactivated when the parent is (the parent
might already be dead).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>