node: fix ref-leak on recursive deactivation
We cannot rely on v_pre as we might recurse into a child, overwriting
v_pre locally on the stack. Therefore, we must remember the 1-bit state
describing whether the node actually ever was active.

Instead of adding node->was_active, we introduce a new active-state
RELEASE_DIRECT, which is taken instead of BIAS if the node was never
active. This allows us to remember 1-bit of state, and thus we can safely
drop the activation-reference after calling ->release_cb().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>