connection: acquire parent ep when adding connections
Instead of checking ep_is_active(), we now properly acquire the endpoint
when adding a connection. This catches races where the endpoint is
deactivated immediately after we checked ep_is_active().

There is no need to acquire the parent bus. The connections are cleaned up
during ep_release(), so its enough to know the ep is active. If the bus
was already inactive, it will deactivate the ep later, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>