test: add -f/--fork to fork even if combined with -t
If we only run a single test, but in a loop, we still want it to fork()
before running each test. Otherwise, our crappy, leaky tests might run out
of resources.

Yeah, we should fix the tests, I guess.. but still, having a fresh
environment for each test is required for stuff like dropping privileges.
So provide -f anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>