metadata: let kdbus_meta_size() modify the attach mask
kdbus_meta_size() and kdbus_meta_write() are always called in pairs,
so we only need to compare the namespace once, and rely on the caller
side to to submit the same set of flags to both calls.

Pass the mask as pointer to kdbus_meta_size() so attach flags can
be dropped on namespace mismatches.

Signed-off-by: Daniel Mack <daniel@zonque.org>