domain: reference mount namespace
Record the mount namespace at the time of domain creation.
We don't take an active reference to it, because

a) there is no API to do so, and
b) we don't have to, as we're only using the pointer to compare
   mount namespaces, not do anything real with it.

In metadata, drop _EXE items if mount or PID namespaces are crossed.

Signed-off-by: Daniel Mack <daniel@zonque.org>