connection: return msg-size in CMD_RECV (ABI break)
An activator usually does not want to map the whole 16MB pool as it never
looks at more than the first connection. Therefore, return the msg-size in
CMD_RECV in combination with the pool offset. This way, the activator can
mmap() just the part it's interested in, use it to log some metadata about
the activation request, unmap it again and start the implementor.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>