metadata: rework for receiver translation
In order to support per-recipient namespace translations, buffers
must be allocated and filled only when we know for which connection
we have to translate. For this, rework the metadata implementation:

* Instead of having a metadata object that carries a blob of data
  which grows over time when more items are attached, only store the
  kernel view of resources internally and don't magically translate
  anything in other namespaces at the time of gathering information.

* In order to get the recipient's view of translated information,
  provide a function called kdbus_meta_export() that returns an
  allocated buffer that can be written to the recipient pool.
  The metadata object is not touched by this operation, hence
  it can be re-used mutliple times.

* Make kdbus_meta ref-counted, so kmsg->meta can take a reference to
  conn_src->owner_meta

* Add a function to fill the database with items from structs, called
  kdbus_meta_fake(). The only use case are connection that provide
  faked credentials at HELLO time. Once this function was called, a
  subsequent kdbus_meta_collect() will be a no-op.

Also, the namespaces are no longer pinned in the domain now, but
in the connection, at the time when the HELLO ioctl is called.

Another reference to the current user_ns is taken in the metadata
object itself, so we can compare namespaces and drop CAPS items if
they differ.

Signed-off-by: Daniel <daniel@zonque.org>