metadata: handle EXE with chroot/mnt-ns
A process can change its root path to any arbitrary value, using chroot or
mnt-namespaces. Therefore, we need to make sure to correctly translate EXE
paths to the target process.

We'd require access to something like __d_path() to do that, but
unfortunately, it is not exported to modules. For now, drop EXE if the
root path differs between connection and target process. If they match,
use d_path().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>