pool: fix tracking user-space refs to slices
Make sure we properly track kernel *and* user-space references to slices.
As we don't want full-blown ref-counters, we leave them as boolean.
However, we now tack both independently, instead of making the kernel-ref
a sub-reference of the public ref.

This fixes a bug where we move messages that are being PEEKed on to an
implementor. The activator would segfault, as the slice is freed. Now,
only the kernel reference is dropped, the user reference will stay until
it is FREEd.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>