metadata: provide module-option to set attach-flags masks
A new module option, called "kdbus.attach_flags_mask", can be used to set
a mask of attach-flags that is applied to the input mask on
kdbus_meta_export(). In other words, attach-flags not listed in that mask
will never be returned to user-space.

We still collect the metadata in the kernel, and we still have all
code-paths enabled. However, missing flags are never handled in any
user-space interactions.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>