endpoint: restrict custom endpoint names to "<uid>-<name>"
Make sure custom endpoint names follow the same restriction as custom
buses. They must be prefixed with "<uid>-".

This will allow adding more custom files to a bus subdirectory later,
without any name clashes with possibly already used endpoints.

Note that the UID *MUST* be provided in the user-namespace of the parent
domain. A domain always has a user-namespace that it is attached to and
*any* accounting for *all* entities on that domain is always done in the
owning user-namespace!

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>