bus: keep sending broadcasts if collecting metadata fails
Currently, we silently drop messages if we cannot collected requested
metadata. This is bad as we tell people that we never drop messages
without telling them.

We have 2 choices: Either we increase the dropped-messages counter, or we
send the message, anyway.

This patch implements method 2. Metadata might be dropped for a lot of
reasons and it is always optional. Metadata never drops the privilege
level of a message, it always raises it! Furthermore, it's the
responsibility of the receiver to properly detect missing metadata. Thus,
it's safe to continue sending messages if we cannot collect requested
metadata.

(do the same for eavesdropping, as it is subject to the same rules)

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>