metadata: fix dead-lock on kdbus_meta_export()
kdbus_meta_export() might lock the target connection so we must not call
it while holding _any_ connection. This might deadlock due to random
locking order.

This patch makes sure to export metadata before we lock any connection.
It's not really a beauty, but it works! We can clean it up later.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>