metadata: preserve INVALID_UID when exporting creds into destination
When exporting creds into destination, if the meta creds were faked then
use the from_kuid() form instead of from_kuid_munged() this way we
preserve the INVALID_UID (uid_t) -1 and do not convert it to nobody.

This is useful for the bus proxy. We can change it later if it happens
this is not what we want. This is just an internal representation of
uid/gid values that can be changed at any time.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>