connection: queue on monitors first
Lets make sure to queue unicast messages on monitors _before_ queuing on
the destination connection. As it turned out, we experienced replies being
queued on monitors before the original message was queued.

Lets make sure we disallow any connection to react to a message _before_
the monitors got the message queued.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>