connection: document why KDBUS_ITEM_NAMES is consistent
Our kdbus_name_lock() call in kdbus_conn_kmsg_send() guarantees that the
whole name registry is read-locked. Therefore, the names a connection
owned cannot change while we send a message.

Add a comment that explains this. We need this behavior to guarantee the
names in KDBUS_ITEM_NAMES are consistent with the names used for message
delivery and policy handling.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>