message.c: consolidate kdbus_msg_scan_items() and kdbus_msg_pin_files()
Revert the split of these two functions. The original idea was to parse
the message again when moving it from one pool to another. Now, with
the two-fold slice logic, that isn't necessary anymore, so the
kdbus_msg_scan_items() can as well pin the files again, and be private
to message.c.

Signed-off-by: Daniel Mack <daniel@zonque.org>