util: kdbus_meta_set_attach_flags() -> kdbus_sanitize_attach_flags()
The kdbus_meta_set_attach_flags() name is really misleading. It suggests
that we set the attach-flags of a meta object, which is really not what
we're doing.

Rename it to kdbus_sanitize_attach_flags() and move it into util.c.
Furthermore, don't inline it as this is really not a critical code-path.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>