message, connection: de-deuplicate struct members
Rather than keeping track of vecs, fds and memfds in both the kmsg and
the queue_entry, factor them out in an ref-counted struct called
'struct kdbus_msg_resources'. Each queue item takes a ref on them
when they are created, and they stay around as long as at least one
message is pending.

Note, however, that the actual payload resides in its own slice, and
the 'src_addr' pointers are not valid once the kmsg is freed. We might
want to add an extra tracking flag for that, so we can BUG_ON() when
src_addr is called, but the SEND ioctl has already returned.

Cleanup up and move code around while at it.

Signed-off-by: Daniel Mack <daniel@zonque.org>