test: add the bus minimal attach flags tests
As it was discussed the kdbus_meta_attach_mask should only be used when
exporting metadata, but the bus creator can request a minimal flags mask
from the connection. So introduce a set of tests that will make sure
that connections providing different send attach masks will work as
expected on buses that require a minimal set of metadata.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>