connection: simplify error-path in kdbus_conn_new()
Lets make sure a connection has all properties set once allocated.
Anything dynamically allocated has proper NULL handling in destructors.
Hence, we can safely use kdbus_conn_unref() in the error-path and
everything will work out just fine.

We have to re-order the initialization a bit to make sure all fields are
properly set. This shouldn't break anything, as it is offline
initialization only. The object is not visible to anyone but our task, so
the order should not matter, except for error-handling.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>