connection: account connection limits to real-user
Accounting of limits and other resources is always done against the real
user, not the fsuid/euid! Take a look at "struct cred", where cred->user
always points to the accounting of cred->uid, not cred->euid/fsuid!

This is important to get right. Even if you change the effective UID, your
actions will always be accounted on your real user. The only thing the
EUID changed, is permission checks!

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>