kdbus.h: report size of all exported pool slices (ABI break)
For each ioctl that reports an offset where returned information can be
found, return a size as well. This field will *always* contain the same
length information that the struct returned in the pool, but it is
available before the mapping is done, and can hence be used to only
map parts of the buffer.

Signed-off-by: Daniel Mack <daniel@zonque.org>