pool: drop kdbus_pool_slice_flush()
With the new iov iter functions, VFS does the dcache page flushing for
us, so there's no need to do it manually.

The only situation in which we have to flush pages ourselves is in the
open-coded pool-to-pool copy loop, but it becomes much easier there as
well, as we're operating on a page basis already.

Signed-off-by: Daniel Mack <daniel@zonque.org>