bus: fix KDBUS_CMD_BUS_CREATOR_INFO ioctl
The iov count in kdbus_cmd_bus_creator_info() is broken some iov will
overwrite others, so fix the count and while we are it re-order some
logic. This way the function will be able later to use the bus owner
attach mask to hide some of its metadata. It will be added later.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>