kdbus.h: add enum kdbus_recv_return_flags
Add an enum for RECV and sync-SEND ioctls, and make the kernel reports
failures in installing file descriptors in a received message.

The message will be received by userspace nevertheless, with return
code 0, but the fds that failed to get installed will have the value -1,
and KDBUS_RECV_RETURN_INCOMPLETE_FDS is set in the reply's return_flags.

Signed-off-by: Daniel Mack <daniel@zonque.org>