util: fix race when copying 'sized' objects from userspace
We copy the 'size' field of userspace objects twice, first to get the
overall size and then again as part of the object. We must properly verify
the size didn't change in between, otherwise we must fail or user-space
might trigger page-faults in the kernel.

Note that we could just avoid copying it twice, but that would mean
hard-coding a lot of helpers like memdup_user() as we have additional
constraints then.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>