kdbus: metadata: collect only the needed connection info
As it is done with kdbus_meta_add_current() add a which argument to
kdbus_meta_add_conn_info(), this way we choose what metadata to collect.

Collecting owned names is a heavy operation, we should not do it if
receivers did not request it. Note that the owned names are not returned
or exported if the receiver did not request them, they are masked, but
even if they are useless we collect them, so just stop doing that and
make sure to collect the only needed information.

kdbus_meta_add_conn_info() collects only KDBUS_ATTACH_NAMES and
KDBUS_ATTACH_CONN_DESCRIPTION, so document this.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>