metadata: kdbus_meta_add_conn_info() minor optimization
Do not take the connection lock if we do not have to, the
conn->description is only writable when the connection is created.

The perfect solution would be to take the connection lock only when
collecting names, and have another per meta lock for owned names and
connection description, since these are the only metadata items that
are updated between conn_info ioctl calls. The lock should be taken
only when doing conn_info() since it is done on the already cached
metadata. Other metadata objects are per message, they are created
and destroyed when sending new messages so they are only valid in the
context of the send ioctl.

Anyway just push this one, later we can improve things.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>