metadata: limit scope of meta->user_namespace
The only reason we pin the user-namespace of the meta owner is to drop
CAPS during _export() if the namespaces differ. This is needed as there is
no nice way to translate caps right now.

Fix the code to limit meta->user_namespace to CAPS only, drop it from
CREDS. Rename it to caps_namespace to not confuse readers.

Note that all this namespace pinning is a HACK! There is really no reason
to ever reuse it for any other metadata. Lets make this clear, and we will
try to drop it once upstream and suitable helpers have been merged.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>