item: drop memzero_explicit()
The case where memzero_explicit() is important is if the local variable
scope of the target is known and limited to the local function. For
kdbus_item_set(), this is really not the case. Use memset() instead.
Furthermore, we copy the data to user-space, so if the compiler drops the
memset(), it'd be totally broken. It cannot optimize writes away if the
data is read and copied later on.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>