item: add struct kdbus_item_header
Never allocate a kdbus_item object on the stack. It's only there as
memory layout definition and can, in theory, grow to arbitrary sizes.

Instead, add a struct kdbus_item_header to item.h that just describes
the header of an item, which is all we care for when manually
assembling an item from kernel space.

Signed-off-by: Daniel Mack <daniel@zonque.org>