test: tree-wide check if all uids and gids are mapped in current userns
Tests that are running in user namespaces where not all the uids and
gids are mapped will be SKIPPED automatically, this is the case of
unprivileged which turn privileged in its userns.

We skip test since even if we have CAP_SETUID we also need the target
uid to be mapped in that userns. Well, we do not need all the uids but
we are just lazy here...

The ones that do not like CONFIG_USER_NS can also test kdbus as it was
in a container, running under root:
  # ./test/kdbus-test --mntns --pidns

With this change unprivileged running on kernel with CONFIR_USER_NS can run
kdbus tests:
$ ./test/kdbus-test --userns --uidns "0 1000 1" --gidns "0 1000 1" --mntns --pidns

(but some tests are skipped)

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>