purge: drop talk cache
Once we use current->creds as policy credentials for UNICASTS, and
file->f_creds for receiver-based policy checks, we cannot use the cache as
we have it now. The creds may change at any time and it's not clear which
creds were used for a given policy check.

Furthermore, the current cache is highly inefficient. It uses only 64
buckets for the whole bus cache. It needs exclusive locking for lookups
and requires _full_ cache traversals whenever a name changes ownership, a
connection dies or a policy is changed.

Drop the cache for now.

Once our code is considered ready and we do benchmarking, we can re-add
it. However, I really think it should look more like this:

    Each *connection* has a policy-cache. This caches the results of the
    respective kdbus_policy_query() calls so we can skip calling into the
    policy (and thus skip locking the policy/bus) on cache hits. On cache
    misses, we call into the policy and cache the result with a back-link
    to the policy-entry that produced this result. Whenever the policy-db
    changes, it drops all cache entries that were generated by a given
    policy-entry.

    As a bonus, our cache can be based on conn->lock. There is no need to
    lock all caches. Furthermore, the policy-db can be updated to use RCU
    as policy-db modifications are definitely not considered a fast-path.
    We then can synchronize on policy-updates and get a lock-free policy
    db.

Anyway, unless someone does some benchmarking, it seems rather useless to
create policy-caches. Lets drop it for now, get everything in place
underneath and re-add it later, if needed (and yeah, chances are high that
some sort of optimization is needed).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>