connection: rename 'reply_count' field to 'request_count'
When reading the code you need always to remember that the reply_count
is not about conting replies, but for the pending request issued by a
connection that are still waiting for a reply. So just rename the
variable nane to reflect its usecase without having to refer to its
kerneldoc.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>